==> Step-1-To download Historical Data (1993-2019).py <==
import pandas as pd
from datetime import datetime

start = datetime.now()
print("Downloading data for Season 2019")
Season_2019 = pd.read_excel(r"https://football-data.co.uk/mmz4281/1920/all-euro-data-2019-2020.xlsx", sheet_name=None)
print("Downloading data for Season 2019 completed successfully")
print("Concatenating data for Season 2019")
Season_2019_c = pd.concat(Season_2019, axis=0, ignore_index=True)
print("Concatenating data for Season 2019 completed successfully")
print("Downloading data for Season 2018")
Season_2018 = pd.read_excel(r"https://football-data.co.uk/mmz4281/1819/all-euro-data-2018-2019.xlsx", sheet_name=None)
print("Downloading data for Season 2018 completed successfully")
print("Concatenating data for Season 2018")
Season_2018_c = pd.concat(Season_2018, axis=0, ignore_index=True)
print("Concatenating data for Season 2018 completed successfully")
print("Downloading data for Season 2017")
Season_2017 = pd.read_excel(r"https://football-data.co.uk/mmz4281/1718/all-euro-data-2017-2018.xlsx", sheet_name=None)
print("Concatenating data for Season 2017")
Season_2017_c = pd.concat(Season_2017, axis=0, ignore_index=True)
print("Concatenating data for Season 2017 completed successfully")
print("Downloading data for Season 2016")
Season_2016 = pd.read_excel(r"https://football-data.co.uk/mmz4281/1617/all-euro-data-2016-2017.xls", sheet_name=None)
print("Concatenating data for Season 2016")
Season_2016_c = pd.concat(Season_2016, axis=0, ignore_index=True)
print("Concatenating data for Season 2016 completed successfully")
print("Downloading data for Season 2015")
Season_2015 = pd.read_excel(r"https://football-data.co.uk/mmz4281/1516/all-euro-data-2015-2016.xls", sheet_name=None)
print("Concatenating data for Season 2015")
Season_2015_c = pd.concat(Season_2015, axis=0, ignore_index=True)
print("Concatenating data for Season 2015 completed successfully")
print("Downloading data for Season 2014")
Season_2014 = pd.read_excel(r"https://football-data.co.uk/mmz4281/1415/all-euro-data-2014-2015.xls", sheet_name=None)
print("Concatenating data for Season 2014")
Season_2014_c = pd.concat(Season_2014, axis=0, ignore_index=True)
print("Concatenating data for Season 2014 completed successfully")
print("Downloading data for Season 2013")
Season_2013 = pd.read_excel(r"https://football-data.co.uk/mmz4281/1314/all-euro-data-2013-2014.xls", sheet_name=None)
print("Concatenating data for Season 2013")
Season_2013_c = pd.concat(Season_2013, axis=0, ignore_index=True)
print("Concatenating data for Season 2013 completed successfully")
print("Downloading data for Season 2012")
Season_2012 = pd.read_excel(r"https://football-data.co.uk/mmz4281/1213/all-euro-data-2012-2013.xls", sheet_name=None)
print("Concatenating data for Season 2012")
Season_2012_c = pd.concat(Season_2012, axis=0, ignore_index=True)
print("Concatenating data for Season 2012 completed successfully")
print("Downloading data for Season 2011")
Season_2011 = pd.read_excel(r"https://football-data.co.uk/mmz4281/1112/all-euro-data-2011-2012.xls", sheet_name=None)
print("Concatenating data for Season 2011")
Season_2011_c = pd.concat(Season_2011, axis=0, ignore_index=True)
print("Concatenating data for Season 2011 completed successfully")
print("Downloading data for Season 2010")
Season_2010 = pd.read_excel(r"https://football-data.co.uk/mmz4281/1011/all-euro-data-2010-2011.xls", sheet_name=None)
print("Concatenating data for Season 2010")
Season_2010_c = pd.concat(Season_2010, axis=0, ignore_index=True)
print("Concatenating data for Season 2010 completed successfully")
print("Downloading data for Season 2009")
Season_2009 = pd.read_excel(r"https://football-data.co.uk/mmz4281/0910/all-euro-data-2009-2010.xls", sheet_name=None)
print("Concatenating data for Season 2009")
Season_2009_c = pd.concat(Season_2009, axis=0, ignore_index=True)
print("Concatenating data for Season 2009 completed successfully")
print("Downloading data for Season 2008")
Season_2008 = pd.read_excel(r"https://football-data.co.uk/mmz4281/0809/all-euro-data-2008-2009.xls", sheet_name=None)
print("Concatenating data for Season 2008")
Season_2008_c = pd.concat(Season_2008, axis=0, ignore_index=True)
print("Concatenating data for Season 2008 completed successfully")
print("Downloading data for Season 2007")
Season_2007 = pd.read_excel(r"https://football-data.co.uk/mmz4281/0708/all-euro-data-2007-2008.xls", sheet_name=None)
print("Concatenating data for Season 2007")
Season_2007_c = pd.concat(Season_2007, axis=0, ignore_index=True)
print("Concatenating data for Season 2007 completed successfully")
print("Downloading data for Season 2006")
Season_2006 = pd.read_excel(r"https://football-data.co.uk/mmz4281/0607/all-euro-data-2006-2007.xls", sheet_name=None)
print("Concatenating data for Season 2006")
Season_2006_c = pd.concat(Season_2006, axis=0, ignore_index=True)
print("Concatenating data for Season 2006 completed successfully")
print("Downloading data for Season 2005")
Season_2005 = pd.read_excel(r"https://football-data.co.uk/mmz4281/0506/all-euro-data-2005-2006.xls", sheet_name=None)
print("Concatenating data for Season 2005")
Season_2005_c = pd.concat(Season_2005, axis=0, ignore_index=True)
print("Concatenating data for Season 2005 completed successfully")
print("Downloading data for Season 2004")
Season_2004 = pd.read_excel(r"https://football-data.co.uk/mmz4281/0405/all-euro-data-2004-2005.xls", sheet_name=None)
print("Concatenating data for Season 2004")
Season_2004_c = pd.concat(Season_2004, axis=0, ignore_index=True)
print("Concatenating data for Season 2004 completed successfully")
print("Downloading data for Season 2003")
Season_2003 = pd.read_excel(r"https://football-data.co.uk/mmz4281/0304/all-euro-data-2003-2004.xls", sheet_name=None)
print("Concatenating data for Season 2003")
Season_2003_c = pd.concat(Season_2003, axis=0, ignore_index=True)
print("Concatenating data for Season 2003 completed successfully")
print("Downloading data for Season 2002")
Season_2002 = pd.read_excel(r"https://football-data.co.uk/mmz4281/0203/all-euro-data-2002-2003.xls", sheet_name=None)
print("Concatenating data for Season 2002")
Season_2002_c = pd.concat(Season_2002, axis=0, ignore_index=True)
print("Concatenating data for Season 2002 completed successfully")
print("Downloading data for Season 2001")
Season_2001 = pd.read_excel(r"https://football-data.co.uk/mmz4281/0102/all-euro-data-2001-2002.xls", sheet_name=None)
print("Concatenating data for Season 2001")
Season_2001_c = pd.concat(Season_2001, axis=0, ignore_index=True)
print("Concatenating data for Season 2001 completed successfully")
print("Downloading data for Season 2000")
Season_2000 = pd.read_excel(r"https://football-data.co.uk/mmz4281/0001/all-euro-data-2000-2001.xls", sheet_name=None)
print("Concatenating data for Season 2000")
Season_2000_c = pd.concat(Season_2000, axis=0, ignore_index=True)
print("Concatenating data for Season 2000 completed successfully")
print("Downloading data for Season 1999")
Season_1999 = pd.read_excel(r"https://football-data.co.uk/mmz4281/9900/all-euro-data-1999-2000.xls", sheet_name=None)
print("Concatenating data for Season 1999")
Season_1999_c = pd.concat(Season_1999, axis=0, ignore_index=True)
print("Concatenating data for Season 1999 completed successfully")
print("Downloading data for Season 1998")
Season_1998 = pd.read_excel(r"https://football-data.co.uk/mmz4281/9899/all-euro-data-1998-1999.xls", sheet_name=None)
print("Concatenating data for Season 1998")
Season_1998_c = pd.concat(Season_1998, axis=0, ignore_index=True)
print("Concatenating data for Season 1998 completed successfully")
print("Downloading data for Season 1997")
Season_1997 = pd.read_excel(r"https://football-data.co.uk/mmz4281/9798/all-euro-data-1997-1998.xls", sheet_name=None)
print("Concatenating data for Season 1997")
Season_1997_c = pd.concat(Season_1997, axis=0, ignore_index=True)
print("Concatenating data for Season 1997 completed successfully")
print("Downloading data for Season 1996")
Season_1996 = pd.read_excel(r"https://football-data.co.uk/mmz4281/9697/all-euro-data-1996-1997.xls", sheet_name=None)
print("Concatenating data for Season 1996")
Season_1996_c = pd.concat(Season_1996, axis=0, ignore_index=True)
print("Concatenating data for Season 1996 completed successfully")
print("Downloading data for Season 1995")
Season_1995 = pd.read_excel(r"https://football-data.co.uk/mmz4281/9596/all-euro-data-1995-1996.xls", sheet_name=None)
print("Concatenating data for Season 1995")
Season_1995_c = pd.concat(Season_1995, axis=0, ignore_index=True)
print("Concatenating data for Season 1995 completed successfully")
print("Downloading data for Season 1994")
Season_1994 = pd.read_excel(r"https://football-data.co.uk/mmz4281/9495/all-euro-data-1994-1995.xls", sheet_name=None)
print("Concatenating data for Season 1994")
Season_1994_c = pd.concat(Season_1994, axis=0, ignore_index=True)
print("Concatenating data for Season 1994 completed successfully")
print("Downloading data for Season 1993")
Season_1993 = pd.read_excel(r"https://football-data.co.uk/mmz4281/9394/all-euro-data-1993-1994.xls", sheet_name=None)
print("Concatenating data for Season 1993")
Season_1993_c = pd.concat(Season_1993, axis=0, ignore_index=True)
print("Concatenating data for Season 1993 completed successfully")
print("Defining list of all seasons")
Season_list = [Season_2020_c,
               Season_2019_c,
               Season_2018_c,
               Season_2017_c,
               Season_2016_c,
               Season_2015_c,
               Season_2014_c,
               Season_2013_c,
               Season_2012_c,
               Season_2011_c,
               Season_2010_c,
               Season_2009_c,
               Season_2008_c,
               Season_2007_c,
               Season_2006_c,
               Season_2005_c,
               Season_2004_c,
               Season_2003_c,
               Season_2002_c,
               Season_2001_c,
               Season_2000_c,
               Season_1999_c,
               Season_1998_c,
               Season_1997_c,
               Season_1996_c,
               Season_1995_c,
               Season_1994_c,
               Season_1993_c,
               ]
print("Concatenating all data into one dataframe")
db_concat = pd.concat(Season_list)
print("Saving csv")
db_concat.to_csv(r'path')
print("File Saved")
print("Success")
end = datetime.now()
time_taken = end - start
print('Time taken to complete: ', time_taken)
